fix: async headers error
