+++ packages/plugin-multi-tenant/src/components/GlobalViewRedirect/index.ts
@@ -1,5 +1,6 @@
 import type { CollectionSlug, ServerProps, ViewTypes } from 'payload'
 
+import { headers as getHeaders } from 'next/headers.js'
 import { redirect } from 'next/navigation.js'
 
 import { getGlobalViewRedirect } from '../../utilities/getGlobalViewRedirect.js'
@@ -16,9 +17,11 @@ export const GlobalViewRedirect = async (args: Args) => {
   const collectionSlug = args?.collectionSlug
 
   if (collectionSlug && args.globalSlugs?.includes(collectionSlug)) {
+    const headers = await getHeaders()
     const redirectRoute = await getGlobalViewRedirect({
       slug: collectionSlug,
       docID: args.docID,
+      headers,
       payload: args.payload,
       tenantFieldName: args.tenantFieldName,
       view: args.viewType,
